--- src/main/java/org/mybatis/dynamic/sql/insert/render/BatchValuePhraseVisitor.java
@@ -1,31 +1,0 @@
-/*
- *    Copyright 2016-2023 the original author or authors.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *       https://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-package org.mybatis.dynamic.sql.insert.render;
-
-import org.mybatis.dynamic.sql.SqlColumn;
-import org.mybatis.dynamic.sql.render.RenderingStrategy;
-
-public class BatchValuePhraseVisitor extends AbstractMultiRowValuePhraseVisitor {
-
-    public BatchValuePhraseVisitor(RenderingStrategy renderingStrategy, String prefix) {
-        super(renderingStrategy, prefix);
-    }
-
-    @Override
-    String calculateJdbcPlaceholder(SqlColumn<?> column) {
-        return column.renderingStrategy().orElse(renderingStrategy).getFormattedJdbcPlaceholder(column, prefix);
-    }
-}
