+++ src/main/java/org/mybatis/dynamic/sql/render/RenderingStrategy.java
@@ -92,16 +92,14 @@ public abstract class RenderingStrategy {
      *
      * @param column column definition used for generating type details in a MyBatis binding. Ignored for Spring.
      * @param prefix parameter prefix used for locating the parameters in a SQL provider object. Typically, will be
-     *     {@link RenderingStrategy#DEFAULT_PARAMETER_PREFIX}. This is ignored for Spring.
-     * @param parameterName name of the parameter. Typically generated by calling
-     *     {@link RenderingStrategy#formatParameterMapKey(AtomicInteger)}
+     *     either "row" or "records[x]" to match the properties of the generated statement object class.
+     * @param parameterName name of the parameter. Typically, this is a property in the row class associated with the
+     *     insert statement.
      * @return the generated binding
      */
     public String getRecordBasedInsertBinding(BindableColumn<?> column, String prefix, String parameterName) {
         return getFormattedJdbcPlaceholder(column, prefix, parameterName);
     }
 
-    public String getMultiRowFormattedJdbcPlaceholder(BindableColumn<?> column, String parameterName) {
-        return getFormattedJdbcPlaceholder(column, parameterName);
-    }
+    public abstract String getRecordBasedInsertBinding(BindableColumn<?> column, String parameterName);
 }
